+++ FrontEnd/login.js
@@ -1,23 +1,8 @@
-import { createElement } from "./element.js";
+import { messageErreurLogin } from "./messageErreur.js";
 
 function afficherAccueil() {
   let pageAccueil = "./index.html";
   location.href = pageAccueil;
-}
-
-function messageErreurLogin() {
-  let spanMessageErreur = document.getElementById("messageErreur");
-
-  if (!spanMessageErreur) {
-    let formConnexion = document.querySelector(".form-connexion");
-    spanMessageErreur = createElement({
-      balise: `span`,
-      id: "messageErreur",
-      text: "l'email ou le mot de passe n'est pas correct",
-      classes: "error",
-    });
-    formConnexion.append(spanMessageErreur);
-  }
 }
 
 function submitLogin() {
@@ -28,7 +13,6 @@ function submitLogin() {
       email: document.querySelector("[name=email]").value,
       password: document.querySelector("[name=mdp]").value,
     };
-    console.log(user);
 
     const reponse = await fetch("http://localhost:5678/api/users/login", {
       method: "POST",
